requirements.txt
EbookLib
customtkinter
packaging
pyinstaller
lxml
google-generativeai
